refactor(#640): rename sheet.py → compose.py — the ADR 0004 layout engine, not the Sheet facade
The stage module holding choose_scale/ViewBlock/StripDepths (the outer
compose-then-pack layout, ADR 0004) was named sheet.py while having
nothing to do with the user-facing Sheet class — a naming trap flagged
by the #635 review. Rename it to compose.py (ADR 0004's own vocabulary)
and update the import sites, comments, CLAUDE.md, and the current-state
module references in ADRs 0004/0005/0008. Public API unchanged:
draftwright.choose_scale still re-exports from __init__.

Frees the sheet.py name for the Sheet facade (next commit).
